Initial commit and program version
This project contains a HTTP client suitable for embedded use, in a
microcontroller on an Internet Of Things device for example. It should
be highly portable and easy to use.

This initial commit contains a working HTTP client that implements a
'GET' requests with chunked/identity encoding, basic authorization, URL
parsing, and resuming a partially completed download (needs testing).
There is still a lot to do to get this into a fully working, production
ready, component, but this initial prototype is good enough.

Things to do:
* Testing things properly
* Implement PUT/DELETE/POST/HEAD/...
* Settle on a proper API
* Fix the TODO comments in the project