logging, DELETE/PUT/..., uninitialized var, ...
There have been many improvements in this commit, and some changes to
how things work. Some deficiencies in the API have been highlighted (but
not resolved at the moment).

The major changes are:
* Logging has been improved, it is more detailed and the functions used
for logging are easier to use. There is also scope for turning logging
off via a few macro definitions if needed.
* Some TODO comments have been removed, others added, it is give and
take at the moment.
* The URL parser has been improved, it still needs more testing, but it
handles all of the unit tests for it.
* DELETE/PUT/HEAD have been implemented. 'PUT' needs more work before it
can actually be used however. 'DELETE' and 'HEAD' should work, although
'DELETE' is less flexible then it could be as it lacks the ability to
generate a body for it (which is uncommon, but allowed).
* The command line interface accommodates the new operations, even if
they are buggy.
* A bug with an uninitialized variable has been fixed in the allocation
routines.
* The program seems to allocate a lot of memory in valgrind, much more
than it should, this needs investigating. It might be some of the system
libraries that are doing it, it might not be however...
* The use of 'atoi' has been removed.

There is still a lot of work to do!