More internal tests, hex chunked encoding, fewer deps
* There are now fewer dependencies on the C standard library; apart from
'assert', 'objdump -t | grep '\*UND\*' is showing; 'strlen', 'memcpy',
'memset', 'memcmp', 'memchr', 'strchr', 'memmove', 'strcmp', and
'strstr'. All of which are easy to implement.
* A bug has been fixed, chunked encoding with hex encoding is now used.
Some function have had to be modified to accomodate this.
* The internal test suite has been improved greatly; various scenarios
can be simulated in pure C without network access by overriding the
callbacks passed into the test suite. Failure cases are currently not
tested.
* The files 'ex1.txt' and 'ex2.txt' have been removed and integrated in
the bash script 't' as here documents, so as not to pollute the
repository with files. The files have also be incorporated in a similar
fashion to the C test suite that has been mentioned.
* Previously read in data is discarded.