Added SSL test code for Unix
* A TLS library has been added as an archive ('tlse.txt'),
called 'tlse', which will probably be removed. It is currently unused,
but is now part of the project history.
* The 'printf' attribute has been added (conditionally on the GCC
compiler being used) so printf format strings are checked.
* For the Unix platform code the platform now works with SSL, no doubt
there are problems with the code (why the hell is SSL so hard to use -
it should not be that difficult, the API is shockingly bad and what is
with the [nearly hundreds] of different ciphers? Pick one and then when
or if it proves to be insecure then release a new version of the
standard the uses a different cipher. I swear intelligence agencies have
deliberately conspired to make SSL and the libraries that implement it
more complex than they need to be so as to introduce vulnerabilities
into everything...).
* The 'Accept-Ranges' field is now handled.
* The callbacks for opening and closing are now passed the operating
system primitives so they can allocate and free memory within
themselves.
* Some dead code has been removed.