New API; internal changes
There is a new API which should be forwards compatible with any changes
needed to save internal state in a HTTP client that can yield back to
the system. Other changes include; adding a millisecond timer function,
reducing stack usage and using a shared buffer for the stack. More of
the standard HTTP operations are also supported, but not well tested.