Added keep-alive option
* A keep-alive option has been added, although cleaning up after it could
be done better and is possible but a little bit complicated. It also has
not been adequately tested. An option for the keep alive has been added
to 'main.c' but there is no easy way to test it.
* White-space has been removed from the end of some lines.
* 'strstr' has been replaced with a case insensitive version.
* All TODO comments have been removed. We are now TODO free. There are
still things that could be improved upon, or changed, or tested, but
those will happen only when needed or requested.