Change assertion to check
An assertion has been changed to a check because the condition can
happen at runtime and is valid for it to happen, this check was making
sure a write of zero length was not attempted, which could not happen
previously but now can as user supplied headers (which can be of zero
length) can be passed to the library.

It is a very minor change.