Renaming many functions, more asserts, comments
There are a lot of small, minor changes, that add up to a lot. Mainly
renaming a lot of functions to include "httpc_" as a prefix, adding
comments and a few more asserts. Also of note is using the "localely.h"
project instead of isdigit or other isX functions and the
toupper/tolower functions.